# docs/swagger.yaml
definitions:
  requestmodels.AddressReq:
    properties:
      addressId:
        type: string
      alternate_phone:
        type: string
      city:
        type: string
      country:
        type: string
      line1:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
      userID:
        type: string
    required:
    - alternate_phone
    - city
    - country
    - line1
    - phone
    - postal_code
    - state
    - street
    type: object
  requestmodels.AdminLoginData:
    properties:
      email:
        type: string
      password:
        minLength: 4
        type: string
    required:
    - email
    - password
    type: object
...
  responsemodels.Response:
    properties:
      after exicution: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
  description: Online Veg Food Delivery.
  title: Go + Gin Veg*n.
  version: "1.0"
paths:
  /address:
    get:
      consumes:
      - application/json
      description: Retrieves all addresses for the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: GetAllAddress
      tags:
      - Address
    post:
      consumes:
      - application/json
      description: Adds a new address for the current user.
      parameters:
      - description: New address data.
        in: body
        name: newAddressData
        required: true
        schema:
          $ref: '#/definitions/requestmodels.AddressReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: AddNewAddress
      tags:
      - Address
  /address/{addressId}:
    patch:
      consumes:
      - application/json
      description: Edits an existing address for the current user.
      parameters:
      - description: The ID of the address to edit.
        in: path
        name: addressId
        required: true
        type: string
      - description: Address edit data.
        in: body
        name: editAddressData
        required: true
        schema:
          $ref: '#/definitions/requestmodels.AddressReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: EditAddress
      tags:
      - Address
...
  /admin/login:
    post:
      consumes:
      - application/json
      description: Logs in an admin user.
      parameters:
      - description: Admin login credentials.
        in: body
        name: loginCredential
        required: true
        schema:
          $ref: '#/definitions/requestmodels.AdminLoginData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responsemodels.Response'
      summary: AdminLogin
      tags:
      - Admin
  /admin/restaurants/pending:
    get:
      consumes:
      - application/json
      description: Retrieves a list of restaurants with a pending status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - AdminRefTokenAuth: []
      summary: PendingRestuarants
      tags:
      - Restaurants
  /admin/restaurants/pending/verify/{id}:
    patch:
      consumes:
      - application/json
      description: Verifies a restaurant by setting its status to "verified".
      parameters:
      - description: The ID of the restaurant to verify.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - AdminRefTokenAuth: []
      summary: VerifyRestaurant
      tags:
      - Restaurants
  /cart:
    get:
      consumes:
      - application/json
      description: Retrieves the cart details for the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: GetCartDetailsOfUser
      tags:
      - Cart
  /cart/{dishid}:
    delete:
      consumes:
      - application/json
      description: Removes a dish from the user's cart.
      parameters:
      - description: The ID of the dish to remove.
        in: path
        name: dishid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: DecrementorRemoveFromCart
      tags:
      - Cart
  /cart/addtocart/{dishid}:
    post:
      consumes:
      - application/json
      description: Adds a dish to the user's cart.
      parameters:
      - description: The ID of the dish to add.
        in: path
        name: dishid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: AddToCart
      tags:
      - Cart
...
securityDefinitions:
  AdminRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
  OtpTempTokenAuth:
    in: header
    name: Authorizations
    type: apiKey
  RestaurantAuthTokenAuth:
    in: header
    name: accesstoken
    type: apiKey
  RestaurantRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
  UserAuthTokenAuth:
    in: header
    name: accesstoken
    type: apiKey
  UserRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
swagger: "2.0"